+++ protobuf/CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add `BuilderInfo` methods to support protoc-plugin 23.0.0. ([#1047])
 * Generalize argument type of `PbList.from` from `List<T>` to `Iterable<T>`.
   ([#1054])
+* Fix clearing oneof fields with `GeneratedMessage.clear`.
 
 [#742]: https://github.com/google/protobuf.dart/pull/742
 [#853]: https://github.com/google/protobuf.dart/pull/853